+++ MovieTicketBooking/src/main/java/com/capg/mtb/repository/IMovieRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.capg.mtb.model.Movie;
 
-public interface IMovieRepository extends JpaRepository<Movie, Long> {
+public interface IMovieRepository extends JpaRepository<Movie, Integer> {
 	@Query(value = "select * from Movies m where theatre_id = 1", nativeQuery = true)
-	List<Movie> findByTheatreId(long theatreId);
+	List<Movie> findByTheatreId(int theatreId);
 }
