+++ MovieTicketBooking/src/main/java/com/capg/mtb/service/impl/IMovieServiceImpl.java
@@ -6,11 +6,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import com.capg.mtb.exceptions.MovieNotFoundException;
 import com.capg.mtb.model.Movie;
 import com.capg.mtb.repository.IMovieRepository;
 import com.capg.mtb.repository.IShowRepository;
 import com.capg.mtb.service.IMovieService;
-import com.cg.mtb.exceptions.MovieNotFoundException;
 
 @Component
 public class IMovieServiceImpl implements IMovieService {
@@ -32,13 +32,13 @@ public class IMovieServiceImpl implements IMovieService {
 	}
 
 	@Override
-	public void removeMovie(long id) throws MovieNotFoundException {
+	public void removeMovie(int id) throws MovieNotFoundException {
 		Movie deleteMovie = iMovieRepository.findById(id)
 				.orElseThrow(() -> new MovieNotFoundException("No movie id is found:" + id));
 		iMovieRepository.delete(deleteMovie);
 	}
 
-	public Movie viewMovie(long id) throws MovieNotFoundException {
+	public Movie viewMovie(int id) throws MovieNotFoundException {
 		return iMovieRepository.findById(id)
 				.orElseThrow(() -> new MovieNotFoundException("No movie id is found:" + id));
 	}
@@ -49,13 +49,13 @@ public class IMovieServiceImpl implements IMovieService {
 	}
 
 	@Override
-	public List<Movie> viewMovieList(long theatreid) {
+	public List<Movie> viewMovieList(int theatreid) {
 		return iMovieRepository.findByTheatreId(theatreid);
 	}
 
 	@Override
 	public List<Movie> viewMovieList(LocalDate date) {
-		 List<Long> movieIds = iShowRepository.findByDate(date);
+		 List<Integer> movieIds = iShowRepository.findByDate(date);
 		 return iMovieRepository.findAllById(movieIds);
 	}
